Crate tera_text_filters

This crate offers filters for the Tera engine that are centered around text transformations.

Functions

camel_case
Converts text into CamelCase.

kebab_case
Converts text into kebab-case.

lower_case
Converts text into lowercase.

mixed_case
Converts text into mixedCase.

register_all
Registers all available filters for a given Tera instance.

snake_case
Converts text into snake_case.

title_case
Converts text into Title Case.

upper_case
Converts text into UPPERCASE.